Added demo to solution
