+++ Kernel.sln
@@ -94,6 +94,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Spruce", "..\XSharp\source\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XSharp.x86", "..\XSharp\source\XSharp.x86\XSharp.x86.csproj", "{7370A62F-12DA-4181-BE3B-009D0926CA7E}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BeepDemo", "BeepDemo", "{83BA8293-9BC0-496E-A9D4-7F8365BEEFCE}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BeepDemo", "Demos\BeepDemo\BeepDemo.csproj", "{5EF63907-CB93-4FA0-BA9C-B73EC7001365}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -335,6 +339,14 @@ Global
 		{7370A62F-12DA-4181-BE3B-009D0926CA7E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7370A62F-12DA-4181-BE3B-009D0926CA7E}.Release|x86.ActiveCfg = Release|Any CPU
 		{7370A62F-12DA-4181-BE3B-009D0926CA7E}.Release|x86.Build.0 = Release|Any CPU
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365}.Debug|x86.Build.0 = Debug|Any CPU
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365}.Release|x86.ActiveCfg = Release|Any CPU
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -380,6 +392,8 @@ Global
 		{E35E0DBF-555F-4D38-8F28-ACDFA9DC97BD} = {5FF9BF2A-5162-4F12-82B6-1693AD776636}
 		{0812DD0A-4CEE-4376-B78A-02EBCBAA14C2} = {3CD3D9A5-9BC5-4FEB-8D63-4D535C0ABB78}
 		{7370A62F-12DA-4181-BE3B-009D0926CA7E} = {3CD3D9A5-9BC5-4FEB-8D63-4D535C0ABB78}
+		{83BA8293-9BC0-496E-A9D4-7F8365BEEFCE} = {B56A6119-1B8F-44E4-9446-291E52F47D4C}
+		{5EF63907-CB93-4FA0-BA9C-B73EC7001365} = {83BA8293-9BC0-496E-A9D4-7F8365BEEFCE}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1A1E8F1D-82B3-471F-9B59-0350DEA9203D}
